+++ Heys/common/io.h
@@ -13,7 +13,7 @@ public:
 
 	std::string getFileName(const std::string& filePath, bool withExtension, char seperator);
 	std::vector<char> readAllBytes(const std::string& filename);
-	int getFileSize(const std::string& from);
+	long long getFileSize(const std::string& from);
 	int getDataBlock(const std::string& from, data_t& to);
 	int setDataBlock(const data_t& from, const std::string& to);
 private:
